+++ crates/zng/src/lib.rs
@@ -594,7 +594,7 @@ mod __prelude {
 
     pub use zng_var::{
         IntoValue, IntoVar, Var, VarValue, const_var, context_var, expr_var, merge_var, var, var_from, var_getter, var_state, when_var,
-    };
+    }; // TODO(breaking) add flat_expr_var!
 
     pub use crate::var::animation::easing;
 
@@ -799,7 +799,7 @@ mod __prelude_wgt {
     pub use zng_var::{
         ContextVar, IntoValue, IntoVar, ResponderVar, ResponseVar, Var, VarCapability, VarHandle, VarHandles, VarValue, const_var,
         context_var, expr_var, impl_from_and_into_var, merge_var, response_done_var, response_var, var, var_getter, var_state, when_var,
-    };
+    }; // TODO(breaking): add flat_expr_var!
 
     pub use zng_layout::{
         context::{DIRECTION_VAR, LAYOUT, LayoutDirection, LayoutMetrics},